# scripts/openapi/openapi_v0.9.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenScan3 API v0.9",
    "version": "0.9"
  },
  "paths": {
    "/cameras/": {
      "get": {
        "tags": [
          "cameras"
        ],
        "summary": "Get Cameras",
        "description": "Get all cameras with their current status\n\nReturns:\n    dict[str, CameraStatusResponse]: A dictionary of camera name to a camera status object",
        "operationId": "get_cameras",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/CameraStatusResponse"
                  },
                  "type": "object",
                  "title": "Response Get Cameras Cameras  Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/cameras/{camera_name}": {
      "get": {
        "tags": [
          "cameras"
        ],
        "summary": "Get Camera",
        "description": "Get a camera with its current status\n\nArgs:\n    camera_name: The name of the camera to get the status of\n\nReturns:\n    CameraStatusResponse: A response object containing the status of the camera",
        "operationId": "get_camera",
        "parameters": [
          {
            "name": "camera_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cameras/{camera_name}/preview": {
      "get": {
        "tags": [
          "cameras"
        ],
        "summary": "Get Preview",
        "description": "Get a camera preview stream in lower resolution\n\nNote: The preview is not rotated by orientation_flag and has to be rotated by client.\n\nArgs:\n    camera_name: The name of the camera to get the preview stream from\n    mode: Either ``stream`` for the MJPEG stream or ``snapshot`` for a single JPEG frame\n    fps: Target frames per second for the stream, clamped between 1 and 50 (only used in stream mode)\n\nReturns:\n    StreamingResponse: A streaming response containing the preview stream",
        "operationId": "get_preview",
        "parameters": [
          {
            "name": "camera_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera Name"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(stream|snapshot)$",
              "default": "stream",
              "title": "Mode"
            }
          },
          {
            "name": "fps",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 50,
              "title": "Fps"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cameras/{camera_name}/photo": {
      "get": {
        "tags": [
          "cameras"
        ],
        "summary": "Get Photo",
        "description": "Get a camera photo\n\nArgs:\n    camera_name: The name of the camera to get the photo from\n\nReturns:\n    Response: A response containing the photo",
        "operationId": "get_photo",
        "parameters": [
          {
            "name": "camera_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera Name"
            }
          },
          {
            "name": "image_format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "jpeg",
                "dng",
                "rgb_array",
                "yuv_array"
              ],
              "type": "string",
              "default": "jpeg",
              "title": "Image Format"
            }
          },
          {
            "name": "with_metadata",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "With Metadata"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cameras/{camera_name}/photo/payload/{payload_id}": {
      "get": {
        "tags": [
          "cameras"
        ],
        "summary": "Get Photo Payload",
        "operationId": "get_photo_payload",
        "parameters": [
          {
            "name": "camera_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera Name"
            }
          },
          {
            "name": "payload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payload Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cameras/{camera_name}/restart": {
      "post": {
        "tags": [
          "cameras"
        ],
        "summary": "Restart Camera",
        "description": "Restart a camera\n\nArgs:\n    camera_name: The name of the camera to restart\n\nReturns:\n    Response: A response containing the status code",
        "operationId": "restart_camera",
        "parameters": [
          {
            "name": "camera_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cameras/{camera_name}/awb-calibration": {
      "post": {
        "tags": [
          "cameras"
        ],
        "summary": "Run automatic white balance calibration and lock the gains.",
        "description": "Expose the camera controller's automatic white balance calibration if available.\n\nArgs:\n    camera_name: Target camera identifier.\n    params: Optional tuning parameters forwarded to the controller implementation.\n\nReturns:\n    AutoCalibrateAwbResponse: Locked gains after the calibration.\n\nRaises:\n    HTTPException: When the controller is busy, unsupported, or calibration fails.",
        "operationId": "auto_calibrate_awb",
        "parameters": [
          {
            "name": "camera_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoCalibrateAwbRequest",
                "default": {
                  "warmup_frames": 12,
                  "stable_frames": 4,
                  "eps": 0.01,
                  "timeout_s": 2.0
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutoCalibrateAwbResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AutoCalibrateAwbRequest": {
        "properties": {
          "warmup_frames": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Warmup Frames",
            "description": "Number of frames to discard before reading AWB metadata.",
            "default": 12
          },
          "stable_frames": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Stable Frames",
            "description": "Consecutive frames that must meet the stability tolerance.",
            "default": 4
          },
          "eps": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Eps",
            "description": "Maximum delta between gain values to consider them stable.",
            "default": 0.01
          },
          "timeout_s": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Timeout S",
            "description": "Maximum time budget for the calibration loop in seconds.",
            "default": 2.0
          }
        },
        "type": "object",
        "title": "AutoCalibrateAwbRequest"
      },
      "AutoCalibrateAwbResponse": {
        "properties": {
          "red_gain": {
            "type": "number",
            "title": "Red Gain"
          },
          "blue_gain": {
            "type": "number",
            "title": "Blue Gain"
          }
        },
        "type": "object",
        "required": [
          "red_gain",
          "blue_gain"
        ],
        "title": "AutoCalibrateAwbResponse"
      },
...
      "CameraStatusResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/CameraType"
          },
          "busy": {
            "type": "boolean",
            "title": "Busy"
          },
          "settings": {
            "$ref": "#/components/schemas/CameraSettings"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "busy",
          "settings"
        ],
        "title": "CameraStatusResponse"
      },
      "CameraType": {
        "type": "string",
        "enum": [
          "gphoto2",
          "linuxpy",
          "picamera2"
        ],
        "title": "CameraType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}